refactor: deleted previous unused implem
